gnrc_sixlowpan_frag: move private RB functions to RB module
