+++ tests/gnrc_sixlowpan_frag/Makefile
@@ -12,7 +12,4 @@ DISABLE_MODULE += auto_init
 # we don't need all this packet buffer space so reduce it a little
 CFLAGS += -DTEST_SUITES -DGNRC_PKTBUF_SIZE=2048
 
-# to be able to include gnrc_sixlowpan_frag-internal `rbuf.h`
-INCLUDES += -I$(RIOTBASE)/sys/net/gnrc/network_layer/sixlowpan/frag/
-
 include $(RIOTBASE)/Makefile.include
